// src/domain/models/device-model.ts
export type DeviceModel = {
  id: string;
  code: string;
  exhibition_description: string;
  un_price: number;
  alias: string;
  images: string;
  detailed_description: string;
  un_in_a_box: number;
  net_weight: number;
  gross_weight: number;
  box_price: number;
  un_avaliable_to_sell: number;
};